requirements.txt
# Core UI
streamlit

# LLMs
google-generativeai
transformers
torch

# Evaluation
evaluate
rouge-score

# Utilities
pandas
python-dotenv
